Move server related files to 'server' folder.
