+++ src/server/ffmpeg.ts
@@ -39,18 +39,18 @@ class FFmpeg implements FFmpegInterface {
       });
       */
       
-      this.process = exec(`ffmpeg -i "${filePath}" -codec copy -movflags +faststart "${path.join(__dirname, `.temp/${roomId}/convert.mp4`)}" -progress - -nostats -loglevel error`, (err, stdout, stderr) => {
+      this.process = exec(`ffmpeg -i "${filePath}" -codec copy -movflags +faststart "${path.join(__dirname, `../.temp/${roomId}/convert.mp4`)}" -progress - -nostats -loglevel error`, (err, stdout, stderr) => {
         if (err) return reject(err);
         this.process = null;
 
-        resolve({ stdout, stderr, mp4Path: path.join(__dirname, `.temp/${roomId}/convert.mp4`) });
+        resolve({ stdout, stderr, mp4Path: path.join(__dirname, `../.temp/${roomId}/convert.mp4`) });
       });
     });
   }
 
   convertVideoToHLS(filePath: string, roomId: string): Promise<any> {
     return new Promise((resolve: any, reject: any) => {
-      this.process = exec(`ffmpeg -i "${filePath}" -codec: copy -start_number 0 -hls_time 10 -hls_list_size 0 -f hls "${path.join(__dirname, `./.streams/${roomId}/index.m3u8`)}"`, (err, stdout, stderr) => {
+      this.process = exec(`ffmpeg -i "${filePath}" -codec: copy -start_number 0 -hls_time 10 -hls_list_size 0 -f hls "${path.join(__dirname, `../.streams/${roomId}/index.m3u8`)}"`, (err, stdout, stderr) => {
         if (err) return reject(err);
         this.process = null;
 
@@ -61,7 +61,7 @@ class FFmpeg implements FFmpegInterface {
 
   extractSubtitles(filePath: string, roomId: string): Promise<any> {
     return new Promise((resolve: any, reject: any) => {
-      this.process = exec(`ffmpeg -i "${filePath}" -map 0:s:0 "${path.join(__dirname, `.streams/${roomId}/subtitles.vtt`)}"`, (err, stdout, stderr) => {
+      this.process = exec(`ffmpeg -i "${filePath}" -map 0:s:0 "${path.join(__dirname, `../.streams/${roomId}/subtitles.vtt`)}"`, (err, stdout, stderr) => {
         if (err) return reject(err);
         this.process = null;
